add TypeORM

+++ src/user/user.controller.ts
@@ -27,8 +27,12 @@ export class UserController {
   }
 
   @Get(':id')
-  findOne(@Param('id', ParseUUIDPipe) id: string) {
-    return this.userService.findOne(id);
+  async findOne(@Param('id', ParseUUIDPipe) id: string) {
+    const user = await this.userService.findOne(id);
+    if (!user) {
+      throw new NotFoundException('User not found');
+    }
+    return user;
   }
 
   @Post()
@@ -46,11 +50,11 @@ export class UserController {
     @Param('id', ParseUUIDPipe) id: string,
     @Body() updatePasswordDto: UpdatePasswordDto,
   ) {
-    const user = this.findOne(id);
+    const user = await this.userService.findOne(id);
     if (!user) {
       throw new NotFoundException('User not found');
     }
-    const currentPassword = this.userService.getCurrentPassword(id);
+    const currentPassword = await this.userService.getCurrentPassword(id);
     const isPasswordValid = await compare(
       updatePasswordDto.oldPassword,
       currentPassword,
@@ -70,7 +74,11 @@ export class UserController {
 
   @Delete(':id')
   @HttpCode(204)
-  remove(@Param('id', ParseUUIDPipe) id: string) {
-    return this.userService.remove(id);
+  async remove(@Param('id', ParseUUIDPipe) id: string) {
+    const user = await this.userService.findOne(id);
+    if (!user) {
+      throw new NotFoundException('User not found');
+    }
+    await this.userService.remove(id);
   }
 }
